Refactoring
int to short, main, evaluation flexibility

+++ Game/AI/include/AlfaBetaAI_sssiie.h
@@ -1,5 +1,9 @@
 
 #pragma once
+///
+// Unused anymore because it's same as AlfaBetaAI_sssi, only different evaluation.
+///
+
 #include "../../include/Player.h"
 #include "../../include/UI.h"
 class AlfaBetaAI_sssiie :
@@ -7,7 +11,7 @@ class AlfaBetaAI_sssiie :
 {
 
 public:
-	AlfaBetaAI_sssiie(Board* board, UI* ui, int teamColor) :Player(teamColor)
+	AlfaBetaAI_sssiie(Board* board, UI* ui, short teamColor) :Player(teamColor)
 	{
 		this->board = board;
 		this->ui = ui;
@@ -23,11 +27,11 @@ private:
 	Board* board;
 	Position oldPos, newPos, arrowPos;
 	AmazonMove move; //holds oldPos and newPos
-	// int** amazons; //amazons, controlled by AI;
+	// short** amazons; //amazons, controlled by AI;
 
-	inline float AlfaBeta(Board* searchBoard, int depth, float alpha, float beta, bool maximizingPlayer);
-	inline float AlfaBetaArrow(Board* searchBoard, int depth, float alpha, float beta, bool maximizingPlayer, AmazonMove bestMove);
-	inline float Evaluate(Board* board, int nextMovingTeamColor);
+	inline float AlfaBeta(Board* searchBoard, short depth, float alpha, float beta, bool maximizingPlayer);
+	inline float AlfaBetaArrow(Board* searchBoard, short depth, float alpha, float beta, bool maximizingPlayer, AmazonMove bestMove);
+	inline float Evaluate(Board* board, short nextMovingTeamColor);
 
 };
