+++ Game/include/Board.h
@@ -15,24 +15,24 @@ public:
     bool isMoveAllowed(Position oldPos, Position newPos);
     void placeArrow(Position pos) { board[pos.x][pos.y] = -1; };
     void undoArrow(Position pos) { board[pos.x][pos.y] = 0; };
-    int getPlayer(Position pos) { return board[pos.x][pos.y]; };
-    int** getAmazons();
-    int** getAmazons(int teamColor);
-    int possibleDirections(Position amazon);
-    int allPossibleDirections(int teamColor);
+    short getPlayer(Position pos) { return board[pos.x][pos.y]; };
+    short** getAmazons();
+    short** getAmazons(short teamColor);
+    short possibleDirections(Position amazon);
+    short allPossibleDirections(short teamColor);
 
 
     //All moves finding is a bit messy, might be nicer if I saves queens in separate arrays too, but for now decided to leave all logic on board array
     //Probably it's still possible to implement it nicer
 
 //    std::vector<AmazonMove> findAllMoves(int teamColor); //find all allowd player amazons moves (from Position[0] to Position[1])
-    std::vector<AmazonMove> findAllMoves(int teamColor); //find all allowd player amazons moves (from Position[0] to Position[1])
+    std::vector<AmazonMove> findAllMoves(short teamColor); //find all allowd player amazons moves (from Position[0] to Position[1])
 //    std::vector<AmazonMove> findAllMovesFrom(Vector2i pos, std::vector<AmazonMove>* moves = NULL);
     std::vector<AmazonMove> findAllMovesFrom(Position pos);
-    int countAllMoves(int teamColor);  // same as finding, but instead counting only (for performance improvement)
-    int countAllMovesFrom(Position pos, int *movesCount = NULL);
+    short countAllMoves(short teamColor);  // same as finding, but instead counting only (for performance improvement)
+    short countAllMovesFrom(Position pos, short*movesCount = NULL);
 
-    bool hasMove(int teamColor);
+    bool hasMove(short teamColor);
     bool hasMove(Position amazon);
 
     void printBoard();
