+++ Game/include/Constants.h
@@ -8,9 +8,9 @@
 // general constants
 namespace c
 {
-    const int BOARD_SIZE = 10;
-    const int AMAZONS = 8;       // total amount of amazons on the board. Should be 8 unless board in board class is changed. must be divisible by 2
-    const int MAX_ARROWS = 92;  // max possible amount of arrows on the board. Amount of board squares - QUEESNS. 92 on a standart 10x10 board
+    const short BOARD_SIZE = 10;
+    const short AMAZONS = 8;       // total amount of amazons on the board. Should be 8 unless board in board class is changed. must be divisible by 2
+    const short MAX_ARROWS = 92;  // max possible amount of arrows on the board. Amount of board squares - QUEESNS. 92 on a standart 10x10 board
 
     //const int BOARD_SIZE = 6;
     //const int AMAZONS = 4;       // total amount of amazons on the board. Should be 8 unless board in board class is changed. must be divisible by 2
@@ -20,7 +20,7 @@ namespace c
     //const int AMAZONS = 4;       // total amount of amazons on the board. Should be 8 unless board in board class is changed. must be divisible by 2
     //const int MAX_ARROWS = 12;  // max possible amount of arrows on the board. Amount of board squares - QUEESNS. 92 on a standart 10x10
 
-    const int MAX_SEARCH_DEPTH = 3;
+    const short MAX_SEARCH_DEPTH = 3;
 } 
 
     
@@ -28,16 +28,16 @@ namespace c
 namespace cUI
 {
     // one squere on board size (width or heigth) in pixels
-    const int SQUARE_SIZE = 60;
+    const short SQUARE_SIZE = 60;
 
-    const int BOARD_WIDTH = 600;  //SQUARE_SIZE*BOARD_SIZE + possible offset (sides etc);
-    const int BOARD_HEIGHT = 600; //SQUARE_SIZE*BOARD_SIZE + possible offset (sides etc);
+    const short BOARD_WIDTH = 600;  //SQUARE_SIZE*BOARD_SIZE + possible offset (sides etc);
+    const short BOARD_HEIGHT = 600; //SQUARE_SIZE*BOARD_SIZE + possible offset (sides etc);
 
-    const int WIN_MSG_WIDTH = 300;
-    const int WIN_MSG_HEIGHT = 50;
-}
+    const short WIN_MSG_WIDTH = 300;
+    const short WIN_MSG_HEIGHT = 50;
+} 
 
-typedef std::array<std::array<int, c::BOARD_SIZE>, c::BOARD_SIZE> BoardArray;
+typedef std::array<std::array<short, c::BOARD_SIZE>, c::BOARD_SIZE> BoardArray;
 
 //TODO is it possible to place initial board into namespace without placing BoardArray in it
 namespace init
